# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Loja ABC LTDA",
        "description": "API para gerenciamento de vendas",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retorna todos os produtos cadastrados",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um novo produto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "description": "Dados do produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Produto A"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entidade não processável",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "O campo name é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Retorna todas as vendas registradas",
                "operationId": "248121b133c730c6860e94383dfe3e12",
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sale"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Cria uma nova venda",
                "operationId": "42cd0cb8dcc906da8e52634de671f7bf",
                "requestBody": {
                    "description": "Dados da venda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "amount",
                                    "products"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Venda de Teste"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "name",
                                                "price",
                                                "amount"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Produto A"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venda criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sale"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entidade não processável",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The products field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/sales/{id}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Retorna uma venda específica",
                "operationId": "b8c1920044f4a939f5bd289d5693854a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sale"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sales"
                ],
                "summary": "Atualiza uma venda existente",
                "operationId": "0b4eb346996f4adbb4a6befe213265d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da venda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "amount"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 9
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venda atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sale"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Exclui uma venda existente",
                "operationId": "59896d964ad2d4af7f6efc82fdad4acf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da venda a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Venda excluída com sucesso"
                    },
                    "404": {
                        "description": "Venda não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Schema para representar um produto",
                "properties": {
                    "id": {
                        "description": "ID do produto",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do produto",
                        "type": "string"
                    },
                    "price": {
                        "description": "Preço do produto",
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "description": "Descrição do produto",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductSale": {
                "properties": {
                    "product_id": {
                        "description": "ID do produto",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do produto",
                        "type": "string"
                    },
                    "price": {
                        "description": "Preço do produto",
                        "type": "number",
                        "format": "float"
                    },
                    "amount": {
                        "description": "Quantidade do produto",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Sale": {
                "properties": {
                    "sale_id": {
                        "description": "ID da venda",
                        "type": "integer"
                    },
                    "amount": {
                        "description": "Valor total da venda",
                        "type": "number",
                        "format": "float"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductSale"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}